+++ src/main/java/com/example/todo/controller/AuthController.java
@@ -3,6 +3,10 @@ package com.example.todo.controller;
 import com.example.todo.common.util.JwtUtil;
 import com.example.todo.common.util.UserRole;
 import com.example.todo.dto.*;
+import com.example.todo.dto.request.LoginRequest;
+import com.example.todo.dto.request.SignUpRequest;
+import com.example.todo.dto.response.AuthorizationResponse;
+import com.example.todo.dto.response.MessageResponse;
 import com.example.todo.service.UserService;
 import jakarta.servlet.http.HttpServletRequest;
 import jakarta.servlet.http.HttpServletResponse;
