--- src/main/java/com/example/todo/dto/CreateCommentRequest.java
@@ -1,7 +1,0 @@
-package com.example.todo.dto;
-
-public record CreateCommentRequest(
-        Long todoId,
-        String content
-) {
-}
