--- src/main/java/com/example/todo/dto/UpdateCommentRequest.java
@@ -1,7 +1,0 @@
-package com.example.todo.dto;
-
-public record UpdateCommentRequest(
-        Long id,
-        String content
-) {
-}
