--- src/main/java/com/example/todo/dto/UpdateTodoRequest.java
@@ -1,9 +1,0 @@
-package com.example.todo.dto;
-
-public record UpdateTodoRequest(
-
-        Long id,
-        String title,
-        String content
-) {
-}
