+++ src/main/java/com/example/todo/dto/request/CreateTodoRequest.java
@@ -1,8 +1,10 @@
-package com.example.todo.dto;
+package com.example.todo.dto.request;
 
 import com.example.todo.domain.Todo;
+import jakarta.validation.constraints.NotEmpty;
 
 public record CreateTodoRequest(
+        @NotEmpty(message = "제목은 필수입니다.")
         String title,
         String content
 ) {
