+++ src/main/java/com/example/todo/dto/request/LoginRequest.java
@@ -1,9 +1,9 @@
-package com.example.todo.dto;
+package com.example.todo.dto.request;
 
 import com.example.todo.validation.Password;
 import com.example.todo.validation.Username;
 
-public record SignUpRequest(
+public record LoginRequest(
         @Username
         String username,
         @Password
