+++ src/main/java/com/example/todo/dto/response/AuthorizationResponse.java
@@ -1,4 +1,4 @@
-package com.example.todo.dto;
+package com.example.todo.dto.response;
 
 public record AuthorizationResponse(
         String accessToken,
