+++ src/main/java/com/example/todo/dto/response/MessageResponse.java
@@ -1,4 +1,4 @@
-package com.example.todo.dto;
+package com.example.todo.dto.response;
 
 public record MessageResponse(
         String status,
