+++ src/main/java/com/example/todo/repository/CommentRepository.java
@@ -4,5 +4,4 @@ import com.example.todo.domain.Comment;
 import org.springframework.data.jpa.repository.JpaRepository;
 
 public interface CommentRepository extends JpaRepository<Comment, Long> {
-    void deleteByIdAndAuthor(long id, String loginCustomerName);
 }
