+++ src/main/java/com/example/todo/service/CommentService.java
@@ -4,8 +4,8 @@ import com.example.todo.common.exception.NotExistException;
 import com.example.todo.domain.Comment;
 import com.example.todo.domain.Todo;
 import com.example.todo.dto.CommentInfo;
-import com.example.todo.dto.CreateCommentRequest;
-import com.example.todo.dto.UpdateCommentRequest;
+import com.example.todo.dto.request.CreateCommentRequest;
+import com.example.todo.dto.request.UpdateCommentRequest;
 import com.example.todo.repository.CommentRepository;
 import lombok.RequiredArgsConstructor;
 import org.springframework.security.core.context.SecurityContextHolder;
