+++ src/main/java/com/example/todo/service/UserService.java
@@ -4,8 +4,8 @@ import com.example.todo.common.util.UserRole;
 import com.example.todo.domain.Authority;
 import com.example.todo.domain.Customer;
 import com.example.todo.dto.CustomerInfo;
-import com.example.todo.dto.LoginRequest;
-import com.example.todo.dto.SignUpRequest;
+import com.example.todo.dto.request.LoginRequest;
+import com.example.todo.dto.request.SignUpRequest;
 import com.example.todo.repository.UserRepository;
 import lombok.RequiredArgsConstructor;
 import org.springframework.security.access.AccessDeniedException;
